+++ src/constants/appVersion.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = "13.331.9";
+export const APP_VERSION = "13.332.0";